1) Fixed bug #19 regarding double click to create an appointment in the DayView 2) Added an eumeration called CalendarViews to replace the constants in the Calendar class 3) Added comments to a bunch of classes 4) Moved the agendaView and monthView classes to their respective packages 5) Changed name of AppointmentWeekPresenceInMonth to AppointmentWidgetParts

+++ branches/version-0.9/gwt-cal/src/main/java/com/bradrydzewski/gwt/calendar/client/dayview/DayView.java
@@ -131,7 +131,7 @@ public class DayView extends CalendarView implements HasSettings {
 		
 		if(widget==null || appointment==null) {
 
-			if (getSettings().getTimeBlockClickNumber() == Click.Single &&
+			if (getSettings().getTimeBlockClickNumber() == Click.Double &&
 					element == dayViewBody.getGrid().gridOverlay.getElement()) {
                 int x = DOM.eventGetClientX(event);
                 int y = DOM.eventGetClientY(event);
